Update Menu.java: add 'url.charCounter' action to menu
Expanding the functionalities of the menu in Menu.java by adding an option to count the frequency of a certain character ('a' in this instance) in the URL. This was achieved by calling the 'charCounter' method from URL.java when option 2 is selected. RESTful APIs have been improved in URL.java by replacing manual operations with usage of newly introduced Writer and Reader classes for optimized I/O operations. This refactor improves code readability and ensures proper resource management with classes implementing AutoCloseable interface.